20250201
package kernel as apk & build image with Makefile

+++ make_rootfs.sh
@@ -1,7 +1,16 @@
 #!/usr/bin/env bash
 
-sudo podman run --rm --privileged loongcr.lcpu.dev/multiarch/archlinux --reset -p yes
-podman run --platform linux/loong64 --name "la99pi-alpine-rootfs" ghcr.io/fzdslr/alpine:3.21 sh -c "echo $(base64 -w 0 ./alpine_init.sh)| base64 -d | sh"
-podman export la99pi-alpine-rootfs > rootfs.tar
-podman rm la99pi-alpine-rootfs
+if [ -f "/proc/sys/fs/binfmt_misc/qemu-loongarch64" ]; then
+    echo "binfmt for qemu-loongarch64 is enabled.";
+else
+    echo "binfmt for qemu-loongarch64 is not enabled.";
+    sudo podman run --rm --privileged loongcr.lcpu.dev/multiarch/archlinux --reset -p yes;
+fi;
 
+container_name=$(podman create --platform linux/loong64 ghcr.io/fzdslr/alpine:3.21 sh -c "echo $(base64 -w 0 ./alpine_init.sh)| base64 -d | sh")
+trap "podman rm $container_name" EXIT;
+
+podman cp kernel/linux-99pi.apk $container_name:/root/
+podman cp kernel/kernel.rsa.pub $container_name:/etc/apk/keys/
+podman start -a $container_name
+podman export $container_name > rootfs.tar
